Add files via upload
A list of suggestions from the given token stream. The given list of tokens reflect a sorted
list of tokens of a text. Each token reflects either a single word or a punctuation mark like :.? A
suggestion is either a single word or a combination of following words (delimited by a single space)
and does not include any stopWord or a single character.